converting to ts build, removing some files
