+++ public/manifest.json
@@ -1,9 +1,9 @@
 {
-  "short_name": "Horticurio",
-  "name": "Horticurio",
+  "short_name": "React App",
+  "name": "Create React App Sample",
   "icons": [
     {
-      "src": "logo.png",
+      "src": "favicon.ico",
       "sizes": "64x64 32x32 24x24 16x16",
       "type": "image/x-icon"
     },
