--- src/Pages/index.js
@@ -1,1 +1,0 @@
-export { About, Browse, Shop, Track };
