+++ src/pages/About.tsx
@@ -1,9 +1,7 @@
 import React from 'react';
-import '../styles.scss';
 import Button from '@material-ui/core/Button';
-import ButtonGroup from '@material-ui/core/ButtonGroup';
 import MuiThemeProvider from '@material-ui/core/styles/MuiThemeProvider';
-import MuiTheme from '../MuiTheme.js';
+import MuiTheme from '../shared/styles/MuiTheme.js';
 
 function About() {
   return (
@@ -12,8 +10,7 @@ function About() {
               <p>
                 Welcome to About!
               </p>
-              <Button target="_blank"
-                rel="noopener noreferrer" color="secondary">hello</Button>
+              <Button color="secondary">hello</Button>
               <a
                 className="link"
                 href="https://reactjs.org"
