+++ src/pages/Browse.tsx
@@ -1,19 +1,16 @@
 import React from 'react';
-import '../styles.scss';
 import Button from '@material-ui/core/Button';
-import ButtonGroup from '@material-ui/core/ButtonGroup';
 import MuiThemeProvider from '@material-ui/core/styles/MuiThemeProvider';
-import MuiTheme from '../MuiTheme.js';
+import MuiTheme from '../shared/styles/MuiTheme.js';
 
-function Shop() {
+function Browse() {
   return (
     <MuiThemeProvider theme={MuiTheme}>
       <div className="app" >
               <p>
-                Welcome to the Shop!
+                Welcome to Browse!
               </p>
-              <Button target="_blank"
-                rel="noopener noreferrer" color="secondary">hello</Button>
+              <Button color="secondary">hello</Button>
               <a
                 className="link"
                 href="https://reactjs.org"
@@ -27,4 +24,4 @@ function Shop() {
   );
 }
 
-export default Shop;
+export default Browse;
